Revert "Alert on number of hosts"
