+++ bin/check-graphite-data.rb
@@ -31,25 +31,99 @@ require 'json'
 require 'open-uri'
 require 'openssl'
 
-require 'sensu-plugins-graphite/graphite_proxy/options'
-require 'sensu-plugins-graphite/graphite_proxy/proxy'
-
 class CheckGraphiteData < Sensu::Plugin::Check::CLI
+  option :target,
+         description: 'Graphite data target',
+         short: '-t TARGET',
+         long: '--target TARGET',
+         required: true
 
-  include SensuPluginsGraphite::GraphiteProxy::Options
+  option :server,
+         description: 'Server host and port',
+         short: '-s SERVER:PORT',
+         long: '--server SERVER:PORT',
+         required: true
+
+  option :username,
+         description: 'username for basic http authentication',
+         short: '-u USERNAME',
+         long: '--user USERNAME',
+         required: false
+
+  option :password,
+         description: 'user password for basic http authentication',
+         short: '-p PASSWORD',
+         long: '--pass PASSWORD',
+         required: false
+
+  option :passfile,
+         description: 'password file path for basic http authentication',
+         short: '-P PASSWORDFILE',
+         long: '--passfile PASSWORDFILE',
+         required: false
+
+  option :warning,
+         description: 'Generate warning if given value is above received value',
+         short: '-w VALUE',
+         long: '--warn VALUE',
+         proc: proc(&:to_f)
+
+  option :critical,
+         description: 'Generate critical if given value is above received value',
+         short: '-c VALUE',
+         long: '--critical VALUE',
+         proc: proc(&:to_f)
 
   option :reset_on_decrease,
-    description: 'Send OK if value has decreased on any values within END-INTERVAL to END',
-    short: '-r INTERVAL',
-    long: '--reset INTERVAL',
-    proc: proc(&:to_i)
+         description: 'Send OK if value has decreased on any values within END-INTERVAL to END',
+         short: '-r INTERVAL',
+         long: '--reset INTERVAL',
+         proc: proc(&:to_i)
 
-  option :allowed_graphite_age, 
+  option :name,
+         description: 'Name used in responses',
+         short: '-n NAME',
+         long: '--name NAME',
+         default: 'graphite check'
+
+  option :allowed_graphite_age,
          description: 'Allowed number of seconds since last data update (default: 60 seconds)',
          short: '-a SECONDS',
          long: '--age SECONDS',
          default: 60,
          proc: proc(&:to_i)
+
+  option :hostname_sub,
+         description: 'Character used to replace periods (.) in hostname (default: _)',
+         short: '-s CHARACTER',
+         long: '--host-sub CHARACTER'
+
+  option :from,
+         description: 'Get samples starting from FROM (default: -10mins)',
+         short: '-f FROM',
+         long: '--from FROM',
+         default: '-10mins'
+
+  option :below,
+         description: 'warnings/critical if values below specified thresholds',
+         short: '-b',
+         long: '--below'
+
+  option :no_ssl_verify,
+         description: 'Do not verify SSL certs',
+         short: '-v',
+         long: '--nosslverify'
+
+  option :help,
+         description: 'Show this message',
+         short: '-h',
+         long: '--help'
+
+  option :auth,
+         description: 'Add an auth token to the HTTP request, in the form of "Name: Value",
+                                     e.g. --auth yourapitokenvaluegoeshere',
+         short: '-a TOKEN',
+         long: '--auth TOKEN'
 
   # Run checks
   def run
@@ -58,19 +132,13 @@ class CheckGraphiteData < Sensu::Plugin::Check::CLI
       exit
     end
 
-    proxy = SensuPluginsGraphite::GraphiteProxy::Proxy.new(config)
-    begin
-      results = proxy.retrieve_data!
-      results.each_pair do |_key, value|
-        @value = value
-        @data = value['data']
-        check_age || check(:critical) || check(:warning)
-      end
-
-      ok("#{name} value okay")
-    rescue SensuPluginsGraphite::GraphiteProxy::ProxyError => e
-      unknown e.message
+    data = retrieve_data
+    data.each_pair do |_key, value|
+      @value = value
+      @data = value['data']
+      check_age || check(:critical) || check(:warning)
     end
+    ok("#{name} value okay")
   end
 
   # name used in responses
@@ -87,12 +155,88 @@ class CheckGraphiteData < Sensu::Plugin::Check::CLI
     end
   end
 
+  def format_url_opts(given_opts)
+    url_opts = {}
+
+    if given_opts[:no_ssl_verify]
+      url_opts[:ssl_verify_mode] = OpenSSL::SSL::VERIFY_NONE
+    end
+
+    if given_opts[:username] && (given_opts[:password] || given_opts[:passfile])
+      if given_opts[:passfile]
+        pass = File.open(given_opts[:passfile]).readline
+      elsif given_opts[:password]
+        pass = given_opts[:password]
+      end
+
+      url_opts[:http_basic_authentication] = [given_opts[:username], pass.chomp]
+    end # we don't have both username and password trying without
+
+    if given_opts[:auth]
+      header = "Bearer #{given_opts[:auth]}"
+      url_opts['Authorization'] = header
+    end
+
+    url_opts
+  end
+
+  def format_output
+    output = {}
+    @json_data.each do |raw|
+      raw['datapoints'].delete_if { |v| v.first.nil? }
+      next if raw['datapoints'].empty?
+      target = raw['target']
+      data = raw['datapoints'].map(&:first)
+      start = raw['datapoints'].first.last
+      dend = raw['datapoints'].last.last
+      step = ((dend - start) / raw['datapoints'].size.to_f).ceil
+      output[target] = { 'target' => target, 'data' => data, 'start' => start, 'end' => dend, 'step' => step }
+    end
+    output
+  end
+
+  # grab data from graphite
+  def retrieve_data
+    # #YELLOW
+    unless @raw_data # rubocop:disable GuardClause
+      begin
+        unless config[:server].start_with?('https://', 'http://')
+          config[:server].prepend('http://')
+        end
+
+        url = "#{config[:server]}/render?format=json&target=#{formatted_target}&from=#{config[:from]}"
+
+        handle = open(url, format_url_opts(config))
+
+        @raw_data = handle.gets
+        if @raw_data == '[]'
+          unknown 'Empty data received from Graphite - metric probably doesn\'t exists'
+        else
+          @json_data = JSON.parse(@raw_data)
+          format_output
+        end
+      rescue OpenURI::HTTPError
+        unknown 'Failed to connect to graphite server'
+      rescue NoMethodError
+        unknown 'No data for time period and/or target'
+      rescue Errno::ECONNREFUSED
+        unknown 'Connection refused when connecting to graphite server'
+      rescue Errno::ECONNRESET
+        unknown 'Connection reset by peer when connecting to graphite server'
+      rescue EOFError
+        unknown 'End of file error when reading from graphite server'
+      rescue => e
+        unknown "An unknown error occured: #{e.inspect}"
+      end
+    end
+  end
+
   # type:: :warning or :critical
   # Return alert if required
   def check(type)
     # #YELLOW
     if config[type] # rubocop:disable GuardClause
-      send(type, "value (#{@data.last}) for #{@value['target']} has passed #{type} threshold (#{config[type]})") if below?(type) || above?(type)
+      send(type, "#{@value['target']} has passed #{type} threshold (#{@data.last})") if below?(type) || above?(type)
     end
   end
 
@@ -117,4 +261,14 @@ class CheckGraphiteData < Sensu::Plugin::Check::CLI
     end
   end
 
+  # Returns formatted target with hostname replacing any $ characters
+  def formatted_target
+    if config[:target].include?('$')
+      require 'socket'
+      @formatted = Socket.gethostbyname(Socket.gethostname).first.gsub('.', config[:hostname_sub] || '_')
+      config[:target].gsub('$', @formatted)
+    else
+      URI.escape config[:target]
+    end
+  end
 end
